requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-adminactions==1.6.0
envparse==0.2.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
unicodecsv==0.14.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==2.4.0
beautifulsoup4==4.9.3
requests==2.25.1